chore(vapp): Add schema/types
