+++ internal/provider/vapp/acl_schema.go
@@ -3,7 +3,6 @@ package vapp
 import (
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/types"
 
 	schemaR "github.com/hashicorp/terraform-plugin-framework/resource/schema"
 
@@ -18,15 +17,6 @@ import (
 	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/provider/common/vapp"
 	"github.com/orange-cloudavenue/terraform-provider-cloudavenue/internal/provider/common/vdc"
 )
-
-type aclResourceModel struct {
-	ID                  types.String `tfsdk:"id"`
-	VDC                 types.String `tfsdk:"vdc"`
-	VAppID              types.String `tfsdk:"vapp_id"`
-	VAppName            types.String `tfsdk:"vapp_name"`
-	EveryoneAccessLevel types.String `tfsdk:"everyone_access_level"`
-	SharedWith          types.Set    `tfsdk:"shared_with"`
-}
 
 /*
 aclSchema
